requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
billiard==3.6.3.0
celery==4.4.7
celery-progress==0.0.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-composite-foreignkey==1.1.0
django-compositekey==1.5.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-filter==2.3.0
django-rest-swagger==2.2.0
django-timezone-field==4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
filelock==3.0.12
flower==0.9.5
future==0.18.2
httpie==2.2.0
humanize==2.5.0
idna==2.10
isort==5.3.2
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
lazy-object-proxy==1.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==7.2.0
pipenv==2020.6.2
prometheus-client==0.8.0
pydevd-pycharm==202.6397.98
Pygments==2.6.1
pylint==2.5.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
redis==3.5.3
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.30
virtualenv-clone==0.5.4
wrapt==1.12.1
